--- flights/.gitignore
@@ -1,5 +1,0 @@
-/target/
-/.classpath
-/.project
-/.idea
-
