--- flights/src/main/java/com/bsd/exampleapp/springboot/flights/other/ErrorDetails.java
@@ -1,30 +1,0 @@
-package com.bsd.exampleapp.springboot.flights.other;
-
-import java.util.Date;
-
-public class ErrorDetails {
-	
-	private Date timestamp;
-	private String message;
-	private String details;
-
-	public ErrorDetails(Date timestamp, String message, String details) {
-	    super();
-	    this.timestamp = timestamp;
-	    this.message = message;
-	    this.details = details;
-	  }
-	
-	public Date getTimestamp() {
-	    return timestamp;
-	  }
-
-	  public String getMessage() {
-	    return message;
-	  }
-
-	  public String getDetails() {
-	    return details;
-	  }
-	  
-}
