--- flights/src/main/resources/templates/error.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-	<meta charset="UTF-8">
-	<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
-	<title>Page not found.</title>
-</head>
-<body>
-	<div class="jumbotron">
-		<h1>Nieoczekiwany błąd! </h1>
-		<p>Nie można wyświetlić strony.</p>
-		<a href="/flights" class="btn btn-info" role="button">Powrót</a>
-	</div>
-</body>
-</html>
