+++ src/main/java/com/bsd/exampleapp/springboot/flights/config/MvcConfig.java
@@ -1,0 +1,16 @@
+package com.bsd.exampleapp.springboot.flights.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.ViewControllerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+@Configuration
+public class MvcConfig implements WebMvcConfigurer {
+
+	@Override
+    public void addViewControllers(ViewControllerRegistry registry) {
+        registry.addViewController("/home").setViewName("home");
+        registry.addViewController("/").setViewName("home");
+    }
+
+}
