+++ src/main/resources/application.properties
@@ -1,0 +1,3 @@
+logging.level.org.springframework 	= DEBUG
+spring.h2.console.enabled 			= true
+server.error.whitelabel.enabled 	= false
