+++ src/main/resources/data.sql
@@ -1,0 +1,4 @@
+insert into pigeon values(pigeon_seq.nextval, 'Gołąb 2');
+insert into pigeon values(pigeon_seq.nextval, 'Gołąb 3');
+insert into pigeon values(pigeon_seq.nextval, 'Pigeon 4');
+insert into pigeon values(pigeon_seq.nextval, 'Gołąb 1');
