+++ src/main/resources/templates/home.html
@@ -1,0 +1,90 @@
+<!DOCTYPE HTML>
+<html>
+<head>
+      <meta charset="UTF-8" />
+      <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
+      <script defer src="https://use.fontawesome.com/releases/v5.3.1/js/all.js"></script>
+      <title>Flights App.</title>
+</head>
+<body>
+
+<div class="jumbotron">
+  <div class="container">
+    <h1>Lot</h1>
+    <p>Start : 2018/09/14 godz 9:00.</p> 
+  </div>
+</div>
+<div class="container">
+  <div class="row col-xs-8">
+  
+	  <h3>Doleciały gołębie :</h3>
+	  
+	  <table id="pigeonsTable" class="table table-bordered table-sm">
+			<thead class="thead-light">
+				<tr>
+					<th class="text-center">Lp</th>
+					<th class="text-center">Gołąb</th>
+					<th></th>
+				</tr>
+			</thead>
+			<tbody>
+			</tbody>
+		</table>
+	  <br>
+	  <button type="button" class="btn btn-info btn-lg" id="btnAdd">Dodaj</button>
+	  
+	  <div class="modal fade" id="modalAdd" role="dialog">
+	    <div class="modal-dialog">
+	    
+	      <div class="modal-content">
+	        <div class="modal-header">
+			  <h5 class="modal-title">Dodaj/edytuj gołębia</h5>
+			  <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+				<span aria-hidden="true">&times;</span>
+			  </button>
+			</div>
+	        <div class="modal-body">
+	          <div class="row col-xs-8">
+					<form id="pigeonForm">
+						<div class="form-group">
+							<label for="name" class="control-label col-xs-2">Gołąb</label>
+							<div class="col-xs-4">
+								<input type="text" class="form-control" id="name" placeholder="Podaj nazwę gołębia" required/>
+							</div>
+						</div>
+							<input type="hidden" class="form-control" id="id"/>
+							<input type="hidden" class="form-control" id="no"/>
+						<button style="margin-bottom:10px" type="submit" class="btn btn-info btn-lg">Zapisz</button>
+						<button style="margin-bottom:10px" type="button" class="btn btn-secondary btn-lg" data-dismiss="modal">Anuluj</button>
+					</form>
+				</div>
+	        </div>
+	        <div class="modal-footer" id="modalFooter">
+	        </div>
+	      </div>
+	      
+	    </div>
+	  </div>
+	
+	<div id="footer">
+	</div>
+	</div>	  
+</div>
+
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js"></script>
+<script src="js/utils.js"></script>
+<script src="js/result.js"></script>
+<script src="js/add.js"></script>
+<script src="js/remove.js"></script>
+<script src="js/update.js"></script>
+
+<script>
+	$(document).ready(function(){
+	    $('[data-toggle="tooltip"]').tooltip(); 
+	});
+</script>
+
+</body>
+</html>
